Quiet a number of ia64 compile errors (not all of which are ia64 specific). Strangely, it seems that  (int) / (size_t) is promoted to some non-int type on ia64, I don't know why, since the result could never become larger than int.

+++ lustre/tests/openunlink.c
@@ -93,7 +93,7 @@ int main(int argc, char **argv)
                 exit(1);
         }
         if (st.st_nlink != 0)
-                fprintf(stderr, "st_nlink = %d\n", st.st_nlink);
+                fprintf(stderr, "st_nlink = %d\n", (int)st.st_nlink);
 
         fprintf(stderr, "reading\n");
         rc = read(fd, buf, strlen(T1) + 1);
